fix: [io]Hidden files on the desktop flash by
Gio created a temporary. file, but it was soon gone. This is because obtaining the file name for the file is invalid, and there is a problem determining whether it is hidden. Use the URL to obtain the file name of the file.

Log: Hidden files on the desktop flash by
Bug: https://pms.uniontech.com/bug-view-200989.html