fix: [io/watcher]The newly created file folder in the/directory was not displayed in a timely manner
Root directory '/', gio monitoring must be“ file://// Converting to a path is "//". Therefore, the file information of the root directory "/" monitored by the monitor is the path starting from "//". Using "//path" to call from LocalFile returns“ file:///0.0.0/path ”Error. So previously, * path!="/" was used here to determine, but this judgment was incorrect, resulting in the path monitored by mtp and ftp (/run/user/1000/gvfs) taking the following branch, so the URL became“ mtp:///path ”）Modify to process'//'before the path. After processing, make the following judgment. If the path is incorrect, reuse the uri

Log: The newly created file folder in the/directory was not displayed in a timely manner
Bug: https://pms.uniontech.com/bug-view-204961.html