fix: [fileinfo]File Manager Crash
When the file manager is closed and the asynchronous thread reads the file attributes in dfmio, a static map is used to convert the string corresponding to the ID. This is because this static map deconstructs and obtains an empty string, causing gio to crash when querying file information assertions. Modify all macros defined by Gio to read properties.

Log: File Manager Crash